+++ src/util/arch/arm/simd_utils.h
@@ -181,9 +181,7 @@ static really_inline m128 set1_2x64(u64a c) {
     return (m128) vdupq_n_u64(c);
 }
 
-static really_inline m128 insert32_m128(m128 in, u32 val, const int imm) {
-    return (m128) vsetq_lane_u32(val, (uint32x4_t)in, imm);
-}
+#define insert32_m128(in, val, imm) ((m128) vsetq_lane_u32(val, (uint32x4_t)in, imm))
 
 static really_inline u32 movd(const m128 in) {
     return vgetq_lane_u32((uint32x4_t) in, 0);
